Folder organization, error fixing, Terrain recreation, Inventory integration
The "Useful Scripts" folder was throwing a bunch of errors because of
FirstPerson code. Went into all the scripts and commented the errors
out. Tried to work on working the inventory system in. Got too late, but
looked at it to get a better understanding. Opened up my Title Screen
and saw my terrain for that got somehow deleted. Tried to recreate it,
then realized we still needed a test procedure, so I wrote that.