Implemented Journal Pages, removed NewEditorTest.cs
Full set of journal pages in assets, starting three pages applied.
Other pages can be added by incrementing the number of elements and
making the last element hold the new page.  NewEditorTest.cs tested if
one object was the same as another, and caused four errors because it
wasn't coded for this project.